Initial Push

+++ FrontEnd/src/app/pages/tracker/tracker.page.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { AuthService } from "../../services/auth.service";
-import { Router } from '@angular/router'
+import { FunctionsToUse } from '../../functions/functions.model'
+import { Router } from '@angular/router';
+import Swal from 'sweetalert2'
 
 @Component({
   selector: 'app-tracker',
@@ -8,11 +10,14 @@ import { Router } from '@angular/router'
   styleUrls: ['./tracker.page.scss'],
 })
 export class TrackerPage implements OnInit {
+  public buttonText = "Finish Service"
+  public functions = new FunctionsToUse()
+
   public noTaskShow = false
 
   public completedTask = false
   public jobsOffered;
-  public onGoingJob;
+  public onGoingJob = [];
 
   constructor(
     private authService: AuthService,
@@ -20,41 +25,42 @@ export class TrackerPage implements OnInit {
   ) { }
 
   ngOnInit() {
-    if (this.jobsOffered == undefined) {
-      this.noTaskShow = true
-    }
-    this.authService.allJobsBeingAccepted({state: "completed"}).subscribe((data) => {
-      this.jobsOffered = data
-      this.onGoingJob = this.jobsOffered.jobs
-    })
+    this.onGoingJob.length = 0
+    this.functions.jobsAccepted(this.authService, {state: "completed"}, this.onGoingJob)
+    this.completedTask = true
+    document.getElementById('completed').style.borderBottom = '2px solid rgb(132, 208, 255)'
+    document.getElementById('going').style.borderBottom = 'none'
+    document.getElementById('reject').style.borderBottom = 'none'
   }
+  
   myOnGoingTask() {
-    this.authService.allJobsBeingAccepted({state: "accept"}).subscribe((data) => {
-      console.log(data)
-      this.jobsOffered = data
-      this.onGoingJob = this.jobsOffered.jobs
-    })
+    this.onGoingJob.length = 0
+    this.functions.onGoingTask(this.authService, {state: "accept", user: this.authService.userIDToken}, this.onGoingJob)
+    this.completedTask = false
   }
   alreadyDoneTask(index) {
-    this.authService.addDataToJobOrders({currentUser: this.authService.userIDToken, state: "completed", jobOffer: this.jobsOffered.jobs[index].data}).subscribe((data) => {
-      this.jobsOffered = data
-      this.onGoingJob = this.jobsOffered.jobs
-      this.deleteItemThroughIndex(index)
-      this.http.navigate(['job-orders'])
+    this.authService.addDataToJobOrders({currentUser: this.authService.userIDToken, state: "completed", jobOffer: this.onGoingJob[index].data}).subscribe((data) => {
+      if(data) {
+        Swal.fire({
+          icon: 'success',
+          title: 'Thank you for using our app!',
+          showConfirmButton: false,
+          timer: 1000
+        })
+        this.deleteItemThroughIndex(index)
+        this.http.navigate(['job-orders'])  
+      }
     })
   }
   completedTasks() {
-    this.authService.allJobsBeingAccepted({state: "completed"}).subscribe((data) => {
-      this.jobsOffered = data
-      this.onGoingJob = this.jobsOffered.jobs
-    })
+    this.onGoingJob.length = 0
+    this.functions.completedTask(this.authService, {state: "completed"}, this.onGoingJob)
+    this.completedTask = true
   }
 
   rejectedTasks() {
-    this.authService.allJobsBeingAccepted({state: "rejected"}).subscribe((data) => {
-      this.jobsOffered = data
-      this.onGoingJob = this.jobsOffered.jobs
-    })
+    this.onGoingJob.length = 0
+    this.functions.rejectedTask(this.authService, {state: "rejected"}, this.onGoingJob)
   }
